docs: fix storybook preview setup
